Issue #326 outsource aamks.py to workers (#352)
* aamks.py moved to redis worker

* fix draft launch

* small changes

* slurm improvements after tests

* types and query fixed

* fixes

* merge dev

---------

+++ gui/projects.php
@@ -114,6 +114,10 @@ function copy_scenario() { # {{{
 	if(in_array($_POST['copy_scenario'], $scenarios, true) && $_POST['copy_scenario'] != 'draft'){
 		$_SESSION['header_err'][]="Scenario '$_POST[copy_scenario]' already exists";
 	} else {
+		if($_SESSION['main']['scenario_name'] == 'draft'){
+			$_SESSION['header_ok'][] = "Launch draft scenario!";
+			return;
+		} 
 		if($_POST['copy_scenario'] == 'draft'){
 			$_SESSION['nn']->query("DELETE FROM scenarios WHERE scenario_name='draft'");
 			$disk_delete=implode("/", array($_SESSION['main']['user_home'], $_SESSION['main']['project_name'], 'draft'));
@@ -134,6 +138,7 @@ function copy_scenario() { # {{{
 					copy($from, $to);
 			}
 			ch_scenario($sid[0]['id']);
+
 			// save conf.json again to update scenario number in the file
 			$conf_file = json_decode(file_get_contents(implode("/", array($new_scenario_directory, "conf.json"))), true);
 			$conf_file['scenario_id'] = $sid[0]['id'];
